kaizntree_backend/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
